requirements.txt
Jinja2
markdown
Pygments
gitpython